V1.1.7.6 Fix a bug when condition is where id = 1000000000000. The old version doesn't support bigint value here.

+++ C#/src/Hubble.Data/Hubble.Core/SFQL/Parse/ParseWhere.cs
@@ -1460,7 +1460,7 @@ namespace Hubble.Core.SFQL.Parse
                                     {
                                         //if the expression like: id = 000
 
-                                        int id = int.Parse(expression.Right[0].Text);
+                                        long id = long.Parse(expression.Right[0].Text);
                                         int docid = _DBProvider.GetDocIdFromDocIdReplaceFieldValue(id);
 
                                         if (docid != int.MinValue)
